+++ src/main/java/com/shutter/springserver/game/GameEventListener.java
@@ -1,5 +1,5 @@
 package com.shutter.springserver.game;
 
 public interface GameEventListener {
-    void onGameFinished(long roomId);
+    void onGameFinished(int roomId);
 }
