+++ src/main/java/com/shutter/springserver/key/UserData.java
@@ -7,10 +7,10 @@ import java.util.List;
 
 public class UserData extends User {
 
-    private long id;
+    private int id;
     private String username;
 
-    public UserData(long id, String username, String email, String password, boolean active, List<GrantedAuthority> authorities) {
+    public UserData(int id, String username, String email, String password, boolean active, List<GrantedAuthority> authorities) {
         super(email, password, active, true, true, true, authorities);
         this.username = username;
         this.id = id;
@@ -20,7 +20,8 @@ public class UserData extends User {
         return username;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
+
 }
